# api/swagger.yaml
swagger: "2.0"

info:
  title: Software Engineering 2 Project
  description: This project is a software implementation of the original mockup realized during the Software Engineering 2 class at the University of Trento.
    Project realized by Cosimo Damiano Persia (173195) and Andrei Catalin Coman (170736).
  version: 1.0.0
schemes:
  - https
host: enigmatic-savannah-48655.herokuapp.com
basePath: /

paths:
  /getMap:
    post:
        summary: Get the path from the destination address to the arrival one.
        description: Send the departure and arrival address. The server will return all the information needed in order to represent the path on the map.
        parameters:
          - name: departure
            in: query
            required: true
            description: Departure address
            type: string
          - name: arrival
            in: query
            required: true
            description: Arrival street
            type: string   
        responses:
          200:
            description: Return a JSON which contains an array of coordinates, departure and arrival location, departure and arrival address, route distance and route duration.
            schema:
              type: object
              properties:
                departureAddress:
                  type: string
                departureLocation:
                  $ref: "#/definitions/Location"
                arrivalAddress:
                  type: string
                arrivalLocation:
                  $ref: "#/definitions/Location"
                routeDistance:
                  type: string
                routeDutation:
                  type: string
                polyline:
                  $ref: "#/definitions/polyline"
          503:
            description: Google API service unavailable.
          449:
            description: The server cannot honour the request because the user has not provided the required information.
  /getUserLocation:
    post:
        summary: Convert coordinates in the corrispondent address.
        description: Send a request containing a pair of coordinates (latitude and longitude) and receive the corrisponding address.
        parameters:
          - name: latitude
            in: query
            required: true
            description: Requester latitude
            type: number
          - name: longitude
            in: query
            required: true
            description: Requester longitude
            type: number  
        responses:
          200:
            description: Return a JSON which contains the address of the coordinates provided within the request.
            schema:
              type: object
              properties:
                address:
                  type: string
          503:
            description: Google API service unavailable.
      

definitions:
    Location:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
    polyline:
      type: array
      items:
        type: object
        properties:
          lat:
            type: number
          lng:
            type: number